requirements.txt
django>=4.2
djangorestframework
django-cors-headers
Pillow
numpy
# Add your ML framework here (e.g., torch, tensorflow) when integrating a real model
